Get2Sail_Back_Django/requirements.txt
action==1.4.4
amqp==2.6.1
api==0.0.7
asgiref==3.5.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.13
django-cors-headers==3.11.0
django-crontab==0.7.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
future==0.18.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
nose==1.3.7
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
serializers==0.2.4
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
zipp==3.8.0
